Abstracts behaviors like drawing, moving, and initializing into a super class.
