+++ Actors/Monster.py
@@ -1,18 +1,8 @@
-from definitions import *
+from Actors.Actor import *
 
-class Monster(arcade.Sprite):
+class Monster(Actor):
     def __init__(self, x, y, color):
-        self.x = x
-        self.y = y
-        self._color = color
-
-    def draw(self):
-        icon = arcade.create_rectangle_filled(self.x, self.y, WIDTH, HEIGHT, self.color)
-        icon.draw()
-
-    def move(self, direction):
-        self.x += direction[0]
-        self.y += direction[1]
+        super().__init__(x, y, color)
 
 class SkullMonster(Monster):
     def __init__(self, x, y):
